pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-note-manager"
version = "0.1.0"
description = "Obsidian Vault automation with Clean Architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pytest>=7.0",
    "google-generativeai>=0.3.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src_v2*"]

[project.scripts]
obsidian = "src_v2.entrypoints.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["src_v2"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__"]
fail_under = 80
